+++ loading_more_list/example/lib/demo/nested_scroll_view_demo.dart
@@ -75,7 +75,7 @@ class _NestedScrollViewDemoState extends State<NestedScrollViewDemo>
         color: Colors.blue,
         pullBackOnRefresh: true,
         onRefresh: onRefresh,
-        //maxDragOffset: 100.0,
+        maxDragOffset: 100.0,
         child: extended.NestedScrollView(
           pinnedHeaderSliverHeightBuilder: () {
             return pinnedHeaderHeight;